max100/requirements.txt
Django
django-extensions
djangorestframework
requests